% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\docType{class}
\name{ApiClient}
\alias{ApiClient}
\title{ApiClient}
\format{
An \code{R6Class} generator object
}
\description{
ApiClient Class
}
\details{
ApiClient Class

Generic API client for OpenAPI client library builds.
OpenAPI generic API client. This client handles the client-
server communication, and is invariant across implementations. Specifics of
the methods and models for each application are generated from the OpenAPI Generator
templates.

NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
Ref: https://openapi-generator.tech
Do not edit the class manually.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{basePath}}{Base url}

\item{\code{userAgent}}{Default user agent}

\item{\code{username}}{Username for HTTP basic authentication}

\item{\code{password}}{Password for HTTP basic authentication}

\item{\code{timeout}}{Default timeout in seconds}

\item{\code{retryStatusCodes}}{vector of status codes to retry}

\item{\code{maxRetryAttempts}}{maximum number of retries for the status codes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ApiClient$new()}}
\item \href{#method-CallApi}{\code{ApiClient$CallApi()}}
\item \href{#method-Execute}{\code{ApiClient$Execute()}}
\item \href{#method-deserialize}{\code{ApiClient$deserialize()}}
\item \href{#method-deserializeObj}{\code{ApiClient$deserializeObj()}}
\item \href{#method-clone}{\code{ApiClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiClient$new(
  basePath = NULL,
  userAgent = NULL,
  defaultHeaders = NULL,
  username = NULL,
  password = NULL,
  apiKeys = NULL,
  accessToken = NULL,
  timeout = NULL,
  retryStatusCodes = NULL,
  maxRetryAttempts = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CallApi"></a>}}
\if{latex}{\out{\hypertarget{method-CallApi}{}}}
\subsection{Method \code{CallApi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiClient$CallApi(url, method, queryParams, headerParams, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Execute"></a>}}
\if{latex}{\out{\hypertarget{method-Execute}{}}}
\subsection{Method \code{Execute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiClient$Execute(url, method, queryParams, headerParams, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deserialize"></a>}}
\if{latex}{\out{\hypertarget{method-deserialize}{}}}
\subsection{Method \code{deserialize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiClient$deserialize(resp, returnType, pkgEnv)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deserializeObj"></a>}}
\if{latex}{\out{\hypertarget{method-deserializeObj}{}}}
\subsection{Method \code{deserializeObj()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiClient$deserializeObj(obj, returnType, pkgEnv)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
