% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{client}
\alias{client}
\title{Creates instance of \code{InfluxDBClient}. Equivalent to \code{InfluxDbClient$new} call().}
\usage{
client(url, token, org, retryOptions = NULL)
}
\arguments{
\item{url}{InfluxDB instance URL}

\item{token}{Authentication token}

\item{org}{Organization name}

\item{retryOptions}{Retry options. Use \code{retry_options} function to create it.
Default is \code{NULL} for no retries. `Use `TRUE` for default retry strategy.}
}
\value{
Instance of \code{InfluxDBClient}
}
\description{
Creates instance of \code{InfluxDBClient}. Equivalent to \code{InfluxDbClient$new} call().
}
\examples{

\dontrun{
client <- client(url = "http://localhost:8086",
                 token = "my-token",
                 org = "my-org")
}
}
