% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_api_client.R
\docType{class}
\name{FluxApiClient}
\alias{FluxApiClient}
\title{FluxApiClient}
\format{
An \code{R6Class} generator object
}
\description{
FluxApiClient Class
}
\section{Super class}{
\code{\link[influxdbclient:ApiClient]{influxdbclient::ApiClient}} -> \code{FluxApiClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-deserialize}{\code{FluxApiClient$deserialize()}}
\item \href{#method-clone}{\code{FluxApiClient$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="influxdbclient" data-topic="ApiClient" data-id="CallApi">}\href{../../influxdbclient/html/ApiClient.html#method-CallApi}{\code{influxdbclient::ApiClient$CallApi()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="influxdbclient" data-topic="ApiClient" data-id="Execute">}\href{../../influxdbclient/html/ApiClient.html#method-Execute}{\code{influxdbclient::ApiClient$Execute()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="influxdbclient" data-topic="ApiClient" data-id="deserializeObj">}\href{../../influxdbclient/html/ApiClient.html#method-deserializeObj}{\code{influxdbclient::ApiClient$deserializeObj()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="influxdbclient" data-topic="ApiClient" data-id="initialize">}\href{../../influxdbclient/html/ApiClient.html#method-initialize}{\code{influxdbclient::ApiClient$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deserialize"></a>}}
\if{latex}{\out{\hypertarget{method-deserialize}{}}}
\subsection{Method \code{deserialize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FluxApiClient$deserialize(resp, returnType, pkgEnv)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FluxApiClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
