% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_api.R
\docType{class}
\name{HealthApi}
\alias{HealthApi}
\title{Health operations}
\format{
An \code{R6Class} generator object
}
\description{
influxdbclient.Health
}
\section{Methods}{

\describe{
\strong{ GetHealth } \emph{ Get the health of an instance }


\itemize{
\item \emph{ @param } zap.trace.span character
\item \emph{ @returnType } \link{HealthCheck} \cr


\item status code : 200 | The instance is healthy

\item return type : HealthCheck 
\item response headers :

\tabular{ll}{
}
\item status code : 503 | The instance is unhealthy

\item return type : HealthCheck 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | Non 2XX error response from server.

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  GetHealth  ####################

library(influxdbclient)
var.zap.trace.span <- '{\"trace_id\":\"1\",\"span_id\":\"1\",\"baggage\":{\"key\":\"value\"}}' # character | OpenTracing span context

#Get the health of an instance
api.instance <- HealthApi$new()

result <- api.instance$GetHealth(zap.trace.span=var.zap.trace.span)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HealthApi$new()}}
\item \href{#method-GetHealth}{\code{HealthApi$GetHealth()}}
\item \href{#method-GetHealthWithHttpInfo}{\code{HealthApi$GetHealthWithHttpInfo()}}
\item \href{#method-clone}{\code{HealthApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HealthApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetHealth"></a>}}
\if{latex}{\out{\hypertarget{method-GetHealth}{}}}
\subsection{Method \code{GetHealth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HealthApi$GetHealth(zap.trace.span = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetHealthWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetHealthWithHttpInfo}{}}}
\subsection{Method \code{GetHealthWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HealthApi$GetHealthWithHttpInfo(zap.trace.span = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HealthApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
