% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_client.R
\docType{class}
\name{InfluxDBClient}
\alias{InfluxDBClient}
\title{InfluxDBClient}
\format{
An \code{R6Class} generator object
}
\description{
InfluxDBClient Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{url}}{Database url}

\item{\code{token}}{Authentication token}

\item{\code{org}}{Organization name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InfluxDBClient$new()}}
\item \href{#method-health}{\code{InfluxDBClient$health()}}
\item \href{#method-query}{\code{InfluxDBClient$query()}}
\item \href{#method-write}{\code{InfluxDBClient$write()}}
\item \href{#method-fromAnnotatedCsv}{\code{InfluxDBClient$fromAnnotatedCsv()}}
\item \href{#method-toLineProtocol}{\code{InfluxDBClient$toLineProtocol()}}
\item \href{#method-clone}{\code{InfluxDBClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfluxDBClient$new(url = NULL, token = NULL, org = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-health"></a>}}
\if{latex}{\out{\hypertarget{method-health}{}}}
\subsection{Method \code{health()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfluxDBClient$health()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query"></a>}}
\if{latex}{\out{\hypertarget{method-query}{}}}
\subsection{Method \code{query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfluxDBClient$query(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write"></a>}}
\if{latex}{\out{\hypertarget{method-write}{}}}
\subsection{Method \code{write()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfluxDBClient$write(x, bucket, precision = "ns", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromAnnotatedCsv"></a>}}
\if{latex}{\out{\hypertarget{method-fromAnnotatedCsv}{}}}
\subsection{Method \code{fromAnnotatedCsv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfluxDBClient$fromAnnotatedCsv(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toLineProtocol"></a>}}
\if{latex}{\out{\hypertarget{method-toLineProtocol}{}}}
\subsection{Method \code{toLineProtocol()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfluxDBClient$toLineProtocol(
  x,
  precision,
  measurementCol = "_measurement",
  tagCols = NULL,
  fieldCols = c(`_field` = "_value"),
  timeCol = "_time"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfluxDBClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
