% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ready_api.R
\docType{class}
\name{ReadyApi}
\alias{ReadyApi}
\title{Ready operations}
\format{
An \code{R6Class} generator object
}
\description{
influxdbclient.Ready
}
\section{Methods}{

\describe{
\strong{ GetReady } \emph{ Get the readiness of an instance at startup }


\itemize{
\item \emph{ @param } zap.trace.span character
\item \emph{ @returnType } \link{Ready} \cr


\item status code : 200 | The instance is ready

\item return type : Ready 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | Non 2XX error response from server.

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  GetReady  ####################

library(influxdbclient)
var.zap.trace.span <- '{\"trace_id\":\"1\",\"span_id\":\"1\",\"baggage\":{\"key\":\"value\"}}' # character | OpenTracing span context

#Get the readiness of an instance at startup
api.instance <- ReadyApi$new()

result <- api.instance$GetReady(zap.trace.span=var.zap.trace.span)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReadyApi$new()}}
\item \href{#method-GetReady}{\code{ReadyApi$GetReady()}}
\item \href{#method-GetReadyWithHttpInfo}{\code{ReadyApi$GetReadyWithHttpInfo()}}
\item \href{#method-clone}{\code{ReadyApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadyApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetReady"></a>}}
\if{latex}{\out{\hypertarget{method-GetReady}{}}}
\subsection{Method \code{GetReady()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadyApi$GetReady(zap.trace.span = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetReadyWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetReadyWithHttpInfo}{}}}
\subsection{Method \code{GetReadyWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadyApi$GetReadyWithHttpInfo(zap.trace.span = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadyApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
