% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retry_options}
\alias{retry_options}
\title{Creates instance of \code{RetryOptions}. Equivalent to \code{RetryOptions$new} call().}
\usage{
retry_options(
  retryJitter = 0,
  retryInterval = 5,
  maxDelay = 125,
  maxRetryTime = 180,
  maxAttempts = 5,
  exponentialBase = 2,
  ...
)
}
\arguments{
\item{retryJitter}{Maximum number of random milliseconds included in delay. Default is \code{0}.}

\item{retryInterval}{First retry delay in seconds. Default is \code{5}.}

\item{maxDelay}{Maximum delay between retries in seconds. Default is \code{125}.}

\item{maxRetryTime}{Maximum time to spend retrying in seconds. Default is \code{180}.}

\item{maxAttempts}{Number of retry attempts. Default is \code{5}.}

\item{exponentialBase}{Base for exponential backoff strategy. Default is \code{2}.}

\item{...}{Optional arguments}
}
\value{
Instance of \code{RetryOptions}
}
\description{
Creates instance of \code{RetryOptions}. Equivalent to \code{RetryOptions$new} call().
}
\examples{

\dontrun{
client <- client(url = "http://localhost:8086",
                 token = "my-token",
                 org = "my-org",
                 retry_options(maxAttempts = 3))
}
}
